Add logging to merges, update some times in report
- TimeAppReport.{price,sum} are now Floats
- Add logPromis for inserting logging in Promise chains
- Set requestTimeout to 600000 ms instead of 60000
- Add TODO for setting bulk size on bulk imports

+++ server/api/timeAppReport/sql/timeAppReport.initialize.sql
@@ -14,8 +14,8 @@ BEGIN
     , [comment] VarChar(MAX) NULL
     , [code] VarChar(255) NULL
     , [quantity] Float NULL
-    , [price] Int NULL
-    , [sum] Int NULL
+    , [price] Float NULL
+    , [sum] Float NULL
     , [employeeId] BigInt NULL
     , [timeAppEmployeeId] BigInt NULL
     , [isUpdated] Bit DEFAULT 1 NULL
